pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-llm"
version = "0.1.0"
description = "Local multi-GGUF pipeline orchestrator"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7,<3",
  "PyYAML>=6.0,<7",
  "psutil>=5.9,<6",
  "streamlit>=1.32,<2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]


[project.scripts]
project-llm = "app.cli:main"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["config*"]
